Committing some old AoDK-related changes
This has all been sitting on my checkout for awhile; it's not *just*
AoDK changes; it looks like this also updates the "mod" to use more
modern PythonSDK constructs.  I'm honestly not 100% sure where exactly
the code is sitting.  I suspect that in order to flip back and forth
between BL2 and AoDK extracts, you'd need to change some stuff in
code.  I don't think the code itself has a way of distinguishing
between them automatically yet.  So that's a concern.  :)  Anyway,
didn't want to leave this stuff just on my own hard drives, so here's
a commit for it.